Make global labels have the same spin style as net/hierarchical labels
For legacy reasons, it stored left and right "spin" as flipped integers in the file format.
But the code handled the flip in multiple locations rather than just doing it on file io.
This change unifies the internal code and does the mapping in the file I/O.

+++ eeschema/dialogs/dialog_edit_sheet_pin.cpp
@@ -81,7 +81,7 @@ bool DIALOG_EDIT_SHEET_PIN::TransferDataToWindow()
     m_textName->SelectAll();
     m_textWidth.SetValue( m_sheetPin->GetTextWidth() );
     m_textHeight.SetValue( m_sheetPin->GetTextHeight() );
-    m_choiceConnectionType->SetSelection( m_sheetPin->GetShape() );
+    m_choiceConnectionType->SetSelection( static_cast<int>( m_sheetPin->GetShape() ) );
 
     return true;
 }
